# tests/GalaxyData.Web.Tests/Hosting/Snapshots/HostTests.TheOpenApiDocumentDescribesTheApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GalaxyData",
    "description": "Browse, query and change data across databases and folders of workbooks. Errors are problem details (RFC 9457) with a code.",
    "version": "v1"
  },
  "paths": {
...
    "/api/query/validate": {
      "post": {
        "tags": [
          "Query"
        ],
        "summary": "What is wrong with a query, placed in its text; the parameters it uses (those without values taken as null), and its columns",
        "operationId": "ValidateQuery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryValidationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/query/explain": {
      "post": {
        "tags": [
          "Query"
        ],
        "summary": "How a query would run, without running it; with a grid, the page the grid would fetch",
        "operationId": "ExplainQuery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryExplainDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/query/execute": {
      "post": {
        "tags": [
          "Query"
        ],
        "summary": "A page of a query's rows, as a grid shows them, with their columns and count when asked for",
        "operationId": "ExecuteQuery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryPageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/query/link": {
      "post": {
        "tags": [
          "Query"
        ],
        "summary": "Where a value of a row leads: the query of those rows, and what a grid can browse for them",
        "operationId": "FollowQueryLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryLinkDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrowseSourceDto": {
        "type": "object",
        "properties": {
          "entity": {
            "type": [
              "null",
              "string"
            ]
          },
          "from": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BrowseFromDto"
              }
            ]
          },
          "navigation": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "DiagnosticDto": {
        "required": [
          "code",
          "severity",
          "message",
          "start",
          "end"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/DiagnosticSeverity"
          },
          "message": {
            "type": "string"
          },
          "start": {
            "type": "integer",
            "format": "int32"
          },
          "end": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "DiagnosticSeverity": {
        "enum": [
          "info",
          "warning",
          "error"
        ]
      },
...
      "ExplainFragmentDto": {
        "required": [
          "source",
          "dialect",
          "sql",
          "parameters",
          "strategy",
          "table",
          "estimatedRows",
          "bindJoinTemplate",
          "bindJoinParameters"
        ],
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "dialect": {
            "type": "string"
          },
          "sql": {
            "type": "string"
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExplainParameter"
            }
          },
          "strategy": {
            "type": "string"
          },
          "table": {
            "type": [
              "null",
              "string"
            ]
          },
          "estimatedRows": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "bindJoinTemplate": {
            "type": [
              "null",
              "string"
            ]
          },
          "bindJoinParameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExplainParameter"
            }
          }
        }
      },
      "ExplainNodeDto": {
        "required": [
          "id",
          "operator",
          "detail",
          "site",
          "columns",
          "estimatedRows",
          "inputs",
          "subqueries"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "operator": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "site": {
            "type": [
              "null",
              "string"
            ]
          },
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estimatedRows": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "inputs": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "subqueries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExplainSubqueryDto"
            }
          }
        }
      },
      "ExplainParameter": {
        "required": [
          "name",
          "type",
          "value"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "ExplainPhase": {
        "required": [
          "name",
          "rules",
          "plan"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "rules": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "plan": {
            "type": "string"
          }
        }
      },
...
      "GridFilterDto": {
        "required": [
          "column",
          "conditions"
        ],
        "type": "object",
        "properties": {
          "column": {
            "type": "string"
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GridConditionDto"
            }
          },
          "any": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "GridSortDto": {
        "required": [
          "column"
        ],
        "type": "object",
        "properties": {
          "column": {
            "type": "string"
          },
          "desc": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "GridStateDto": {
        "type": "object",
        "properties": {
          "filters": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/GridFilterDto"
            }
          },
          "where": {
            "maxLength": 10000,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "sort": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/GridSortDto"
            }
          },
          "offset": {
            "maximum": 9223372036854776000,
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          },
          "limit": {
            "maximum": 100000,
            "minimum": 1,
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProblemDetails": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string",
            "description": "What the problem is, for clients to tell problems apart by"
          },
          "traceId": {
            "type": "string",
            "description": "The request's trace, to find it in the logs"
          }
        }
      },
...
      "QueryColumnDto": {
        "required": [
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/TypeDto"
          }
        }
      },
      "QueryExplainDto": {
        "required": [
          "queryText",
          "parameters",
          "summary",
          "diagnostics",
          "schema",
          "plan",
          "nodes",
          "fragments",
          "mergeSql",
          "mergeParameters",
          "phases",
          "text"
        ],
        "type": "object",
        "properties": {
          "queryText": {
            "type": "string"
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryParameterDto"
            }
          },
          "summary": {
            "type": "string"
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticDto"
            }
          },
          "schema": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResultSchemaDto"
              }
            ]
          },
          "plan": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExplainNodeDto"
            }
          },
          "fragments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExplainFragmentDto"
            }
          },
          "mergeSql": {
            "type": [
              "null",
              "string"
            ]
          },
          "mergeParameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExplainParameter"
            }
          },
          "phases": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ExplainPhase"
            }
          },
          "text": {
            "type": "string"
          }
        }
      },
      "QueryExplainRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "parameters": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/QueryParameterInput"
            }
          },
          "grid": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GridStateDto"
              }
            ]
          },
          "verbose": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "QueryLinkDto": {
        "required": [
          "queryText",
          "parameters",
          "browse",
          "key"
        ],
        "type": "object",
        "properties": {
          "queryText": {
            "type": [
              "null",
              "string"
            ]
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryParameterDto"
            }
          },
          "browse": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BrowseSourceDto"
              }
            ]
          },
          "key": {
            "type": [
              "null",
              "array"
            ]
          }
        }
      },
      "QueryLinkRequest": {
        "required": [
          "text",
          "row"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "row": {
            "type": "array"
          },
          "parameters": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/QueryParameterInput"
            }
          },
          "column": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "related": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "catalogVersion": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "QueryPageDto": {
        "required": [
          "queryText",
          "parameters",
          "schema",
          "rows",
          "offset",
          "hasMore",
          "total",
          "stats",
          "warnings"
        ],
        "type": "object",
        "properties": {
          "queryText": {
            "type": "string"
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryParameterDto"
            }
          },
          "schema": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResultSchemaDto"
              }
            ]
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultRowDto"
            }
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "hasMore": {
            "type": "boolean"
          },
          "total": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "stats": {
            "$ref": "#/components/schemas/QueryStatsDto"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticDto"
            }
          }
        }
      },
      "QueryPageRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "parameters": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/QueryParameterInput"
            }
          },
          "grid": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GridStateDto"
              }
            ]
          },
          "includeSchema": {
            "type": "boolean",
            "default": true
          },
          "includeCount": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "QueryParameterDto": {
        "required": [
          "name",
          "type",
          "value"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "value": { }
        }
      },
      "QueryParameterInput": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "value": { }
        }
      },
      "QueryStatsDto": {
        "required": [
          "elapsedMs",
          "fetchedRows",
          "keysSent",
          "fragments"
        ],
        "type": "object",
        "properties": {
          "elapsedMs": {
            "type": "number",
            "format": "double"
          },
          "fetchedRows": {
            "type": "integer",
            "format": "int64"
          },
          "keysSent": {
            "type": "integer",
            "format": "int64"
          },
          "fragments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FragmentStatsDto"
            }
          }
        }
      },
      "QueryTextRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "parameters": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/QueryParameterInput"
            }
          }
        }
      },
      "QueryValidationDto": {
        "required": [
          "success",
          "complete",
          "diagnostics",
          "parameters",
          "columns"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "complete": {
            "type": "boolean"
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticDto"
            }
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsedParameterDto"
            }
          },
          "columns": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/QueryColumnDto"
            }
          }
        }
      },
...
      "ResultRowDto": {
        "required": [
          "id",
          "v"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": [
              "null",
              "string"
            ]
          },
          "v": {
            "type": "array"
          }
        }
      },
      "ResultSchemaDto": {
        "required": [
          "columns",
          "rowIdentity"
        ],
        "type": "object",
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultColumnDto"
            }
          },
          "rowIdentity": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RowIdentityDto"
              }
            ]
          }
        }
      },
...
      "TypeDto": {
        "required": [
          "kind",
          "nullable",
          "text"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ScalarKind"
          },
          "nullable": {
            "type": "boolean"
          },
          "text": {
            "type": "string"
          },
          "precision": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "scale": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "length": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "isAnsi": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
      "UsedParameterDto": {
        "required": [
          "name",
          "given",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "given": {
            "type": "boolean"
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}